requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.4
tqdm==4.66.1
python-dotenv==1.0.0
networkx==3.1
hdbscan==0.8.32
ujson==5.8.0
requests==2.34.0
